# 8308479: [s390x] Implement alternative fast-locking scheme