--- a/src/hotspot/cpu/s390/c1_MacroAssembler_s390.cpp
+++ b/src/hotspot/cpu/s390/c1_MacroAssembler_s390.cpp
@@ -83,60 +83,63 @@ void C1_MacroAssembler::verified_entry(bool breakAtEntry) {
   if (breakAtEntry) z_illtrap(0xC1);
 }
 
-void C1_MacroAssembler::lock_object(Register hdr, Register obj, Register disp_hdr, Label& slow_case) {
+void C1_MacroAssembler::lock_object(Register Rmark, Register Roop, Register Rbox, Label& slow_case) {
   const int hdr_offset = oopDesc::mark_offset_in_bytes();
-  assert_different_registers(hdr, obj, disp_hdr);
 
-  verify_oop(obj, FILE_AND_LINE);
+  const Register tmp   = Z_R1_scratch;
+
+  assert_different_registers(Rmark, Roop, Rbox, tmp);
+
+  verify_oop(Roop, FILE_AND_LINE);
 
   // Load object header.
-  z_lg(hdr, Address(obj, hdr_offset));
+  z_lg(Rmark, Address(Roop, hdr_offset));
 
   // Save object being locked into the BasicObjectLock...
-  z_stg(obj, Address(disp_hdr, BasicObjectLock::obj_offset()));
+  z_stg(Roop, Address(Rbox, BasicObjectLock::obj_offset()));
 
   if (DiagnoseSyncOnValueBasedClasses != 0) {
-    load_klass(Z_R1_scratch, obj);
-    testbit(Address(Z_R1_scratch, Klass::access_flags_offset()), exact_log2(JVM_ACC_IS_VALUE_BASED_CLASS));
+    load_klass(tmp, Roop);
+    testbit(Address(tmp, Klass::access_flags_offset()), exact_log2(JVM_ACC_IS_VALUE_BASED_CLASS));
     z_btrue(slow_case);
   }
 
   assert(LockingMode != LM_MONITOR, "LM_MONITOR is already handled, by emit_lock()");
 
   if (LockingMode == LM_LIGHTWEIGHT) {
-    Unimplemented();
+    fast_lock(Roop, Rmark, tmp, slow_case);
   } else if (LockingMode == LM_LEGACY) {
     NearLabel done;
     // and mark it as unlocked.
-    z_oill(hdr, markWord::unlocked_value);
+    z_oill(Rmark, markWord::unlocked_value);
     // Save unlocked object header into the displaced header location on the stack.
-    z_stg(hdr, Address(disp_hdr, (intptr_t) 0));
+    z_stg(Rmark, Address(Rbox, BasicLock::displaced_header_offset_in_bytes()));
     // Test if object header is still the same (i.e. unlocked), and if so, store the
     // displaced header address in the object header. If it is not the same, get the
     // object header instead.
-    z_csg(hdr, disp_hdr, hdr_offset, obj);
+    z_csg(Rmark, Rbox, hdr_offset, Roop);
     // If the object header was the same, we're done.
     branch_optimized(Assembler::bcondEqual, done);
-    // If the object header was not the same, it is now in the hdr register.
+    // If the object header was not the same, it is now in the Rmark register.
     // => Test if it is a stack pointer into the same stack (recursive locking), i.e.:
     //
-    // 1) (hdr & markWord::lock_mask_in_place) == 0
-    // 2) rsp <= hdr
-    // 3) hdr <= rsp + page_size
+    // 1) (Rmark & markWord::lock_mask_in_place) == 0
+    // 2) rsp <= Rmark
+    // 3) Rmark <= rsp + page_size
     //
     // These 3 tests can be done by evaluating the following expression:
     //
-    // (hdr - Z_SP) & (~(page_size-1) | markWord::lock_mask_in_place)
+    // (Rmark - Z_SP) & (~(page_size-1) | markWord::lock_mask_in_place)
     //
     // assuming both the stack pointer and page_size have their least
     // significant 2 bits cleared and page_size is a power of 2
-    z_sgr(hdr, Z_SP);
+    z_sgr(Rmark, Z_SP);
 
     load_const_optimized(Z_R0_scratch, (~(os::vm_page_size() - 1) | markWord::lock_mask_in_place));
-    z_ngr(hdr, Z_R0_scratch); // AND sets CC (result eq/ne 0).
+    z_ngr(Rmark, Z_R0_scratch); // AND sets CC (result eq/ne 0).
     // For recursive locking, the result is zero. => Save it in the displaced header
-    // location (null in the displaced hdr location indicates recursive locking).
-    z_stg(hdr, Address(disp_hdr, (intptr_t) 0));
+    // location (null in the displaced Rmark location indicates recursive locking).
+    z_stg(Rmark, Address(Rbox, BasicLock::displaced_header_offset_in_bytes()));
     // Otherwise we don't care about the result and handle locking via runtime call.
     branch_optimized(Assembler::bcondNotZero, slow_case);
     // done
@@ -144,35 +147,41 @@ void C1_MacroAssembler::lock_object(Register hdr, Register obj, Register disp_hd
   }
 }
 
-void C1_MacroAssembler::unlock_object(Register hdr, Register obj, Register disp_hdr, Label& slow_case) {
-  const int aligned_mask = BytesPerWord -1;
+void C1_MacroAssembler::unlock_object(Register Rmark, Register Roop, Register Rbox, Label& slow_case) {
   const int hdr_offset = oopDesc::mark_offset_in_bytes();
-  assert_different_registers(hdr, obj, disp_hdr);
+
+  assert_different_registers(Rmark, Roop, Rbox);
+
   NearLabel done;
 
   if (LockingMode != LM_LIGHTWEIGHT) {
     // Load displaced header.
-    z_ltg(hdr, Address(disp_hdr, (intptr_t) 0));
-    // If the loaded hdr is null we had recursive locking, and we are done.
+    z_ltg(Rmark, Address(Rbox, BasicLock::displaced_header_offset_in_bytes()));
+    // If the loaded Rmark is null we had recursive locking, and we are done.
     z_bre(done);
   }
 
   // Load object.
-  z_lg(obj, Address(disp_hdr, BasicObjectLock::obj_offset()));
-  verify_oop(obj, FILE_AND_LINE);
+  z_lg(Roop, Address(Rbox, BasicObjectLock::obj_offset()));
+  verify_oop(Roop, FILE_AND_LINE);
 
   if (LockingMode == LM_LIGHTWEIGHT) {
-    Unimplemented();
-  } else {
+    const Register tmp = Z_R1_scratch;
+    z_lg(Rmark, Address(Roop, hdr_offset));
+    z_lgr(tmp, Rmark);
+    z_nill(tmp, markWord::monitor_value);
+    z_brnz(slow_case);
+    fast_unlock(Roop, Rmark, tmp, slow_case);
+  } else if (LockingMode == LM_LEGACY) {
     // Test if object header is pointing to the displaced header, and if so, restore
     // the displaced header in the object. If the object header is not pointing to
     // the displaced header, get the object header instead.
-    z_csg(disp_hdr, hdr, hdr_offset, obj);
+    z_csg(Rbox, Rmark, hdr_offset, Roop);
     // If the object header was not pointing to the displaced header,
     // we do unlocking via runtime call.
     branch_optimized(Assembler::bcondNotEqual, slow_case);
-    // done
   }
+  // done
   bind(done);
 }