--- a/src/hotspot/cpu/s390/c1_MacroAssembler_s390.hpp
+++ b/src/hotspot/cpu/s390/c1_MacroAssembler_s390.hpp
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2016 SAP SE. All rights reserved.
+ * Copyright (c) 2016, 2023 SAP SE. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -41,18 +41,18 @@
   void initialize_body(Register objectFields, Register len_in_bytes, Register Rzero);
 
   // locking
-  // hdr     : Used to hold locked markWord to be CASed into obj, contents destroyed.
-  // obj     : Must point to the object to lock, contents preserved.
-  // disp_hdr: Must point to the displaced header location, contents preserved.
-  // Returns code offset at which to add null check debug information.
-  void lock_object(Register hdr, Register obj, Register disp_hdr, Label& slow_case);
+  // Rmark        : Used to hold locked markWord to be CASed into obj, contents destroyed.
+  // Roop         : Must point to the object to lock, contents preserved.
+  // Rbox         : Must point to the displaced header location, contents preserved.
+  // Z_R1_scratch : Used as temp and will be killed
+  void lock_object(Register Rmark, Register Roop, Register Rbox, Label& slow_case);
 
   // unlocking
-  // hdr     : Used to hold original markWord to be CASed back into obj, contents destroyed.
-  // obj     : Must point to the object to lock, contents preserved.
-  // disp_hdr: Must point to the displaced header location, contents destroyed.
-  // Z_R0_scratch will be killed as well
-  void unlock_object(Register hdr, Register obj, Register lock, Label& slow_case);
+  // Rmark        : Used to hold original markWord to be CASed back into obj, contents destroyed.
+  // Roop         : Must point to the object to lock, contents preserved.
+  // Rbox         : Must point to the displaced header location, contents destroyed.
+  // Z_R1_scratch : Used as temp and will be killed
+  void unlock_object(Register Rmark, Register Roop, Register Rbox, Label& slow_case);
 
   void initialize_object(
     Register obj,                      // result: Pointer to object after successful allocation.