--- a/src/hotspot/cpu/s390/macroAssembler_s390.cpp
+++ b/src/hotspot/cpu/s390/macroAssembler_s390.cpp
@@ -3261,6 +3261,8 @@ void MacroAssembler::compiler_fast_unlock_object(Register oop, Register box, Reg
   // The object has an existing monitor iff (mark & monitor_value) != 0.
   z_lg(currentHeader, hdr_offset, oop);
   guarantee(Immediate::is_uimm16(markWord::monitor_value), "must be half-word");
+  if (LockingMode == LM_LIGHTWEIGHT)
+    z_lgr(temp, currentHeader);
   z_nill(currentHeader, markWord::monitor_value);
   z_brne(object_has_monitor);
 
@@ -3278,8 +3280,9 @@ void MacroAssembler::compiler_fast_unlock_object(Register oop, Register box, Reg
   } else {
     assert(LockingMode == LM_LIGHTWEIGHT, "must be");
 
-    // don't load currentHead again after monitor check, as it is possible
+    // don't load currentHead again from stack-top after monitor check, as it is possible
     // some other thread modified it.
+    z_lgr(currentHeader, temp);
     fast_unlock(oop, currentHeader, temp, done);
     z_bru(done);
   }