… src/hotspot/cpu/s390/sharedRuntime_s390.cpp
… src/hotspot/share/runtime/arguments.cpp