This PR implements new fast-locking scheme for s390x. Additionally few parameters have been renamed to be in sync with PPC.

Testing done (for release, fastdebug and slowdebug build):
All `test/jdk/java/util/concurrent` test with parameters:
* LockingMode=2 
* LockingMode=2 with -Xint
* LockingMode=2 with -XX:TieredStopAtLevel=1
* LockingMode=2 with -XX:-TieredCompilation

Result is consistently similar to Aarch(MacOS) and PPC, All of 124 tests are passing except `MapLoops.java` because in the 2nd part for this testcase, jvm starts with `HeavyMonitors` which conflict with `LockingMode=2`

BenchMark Result for Renaissance-jmh:

| Benchmark                                | Without fastLock (ms/op) | With fastLock (ms/op) | Improvement |
|------------------------------------------|-------------------------|----------------------|-------------|
| o.r.actors.JmhAkkaUct.runOperation        | 1565.080                | 1365.877             | 12.70%      |
| o.r.actors.JmhReactors.runOperation       | 9316.972                | 10592.982            | -13.70%     |
| o.r.jdk.concurrent.JmhFjKmeans.runOperation | 1257.183                | 1235.530             | 1.73%       |
| o.r.jdk.concurrent.JmhFutureGenetic.runOperation | 1925.158                | 2073.066             | -7.69%      |
| o.r.jdk.streams.JmhParMnemonics.runOperation | 2746.664                | 2836.085             | -3.24%      |
| o.r.jdk.streams.JmhScrabble.runOperation   | 76.774                  | 74.239               | 3.31%       |
| o.r.rx.JmhRxScrabble.runOperation          | 162.270                 | 167.061              | -2.96%      |
| o.r.scala.sat.JmhScalaDoku.runOperation    | 3333.711                | 3271.078             | 1.88%       |
| o.r.scala.stdlib.JmhScalaKmeans.runOperation | 182.746                 | 182.153              | 0.33%       |
| o.r.scala.stm.JmhPhilosophers.runOperation | 15003.329               | 13396.921            | 10.57%      |
| o.r.scala.stm.JmhScalaStmBench7.runOperation | 1669.090                | 1579.900             | 5.34%       |
| o.r.twitter.finagle.JmhFinagleChirper.runOperation | 9601.963                | 10034.404            | -4.52%      |
| o.r.twitter.finagle.JmhFinagleHttp.runOperation | 4403.725                | 4746.707             | -7.79%      |


DaCapo Benchmark Result:

| Benchmark                | Without fast lock (msec) | With fast lock (msec) | Improvement |
|--------------------------|-------------------------|----------------------|-------------|
| DaCapo 9.12 h2 Run 1     | 117,010                 | 108,699              | 7.10%       |
| DaCapo 9.12 h2 Run 2     | 111,986                 | 107,373              | 4.11%       |

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8308479](https://bugs.openjdk.org/browse/JDK-8308479): [s390x] Implement alternative fast-locking scheme (**Enhancement** - P4)


### Reviewers
 * [Lutz Schmidt](https://openjdk.org/census#lucy) (@RealLucy - **Reviewer**)
 * [Martin Doerr](https://openjdk.org/census#mdoerr) (@TheRealMDoerr - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/14414/head:pull/14414` \
`$ git checkout pull/14414`

Update a local copy of the PR: \
`$ git checkout pull/14414` \
`$ git pull https://git.openjdk.org/jdk.git pull/14414/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 14414`

View PR using the GUI difftool: \
`$ git pr show -t 14414`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/14414.diff">https://git.openjdk.org/jdk/pull/14414.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/14414#issuecomment-1587115712)