@offamitkumar 2023-06-22
Hi @RealLucy,

Please check benchmark result in description.