@TheRealMDoerr 2023-06-22
Are the benchmark results stable or do they have a large variance?
Did you also compare against the version without your changes? That should be done for sanity checking. You have modified the code for the legacy mode, too.
I'm not sure what you mean by "With/Without fastLock"? Patch applied? LockingMode selection?