@offamitkumar 2023-06-23
>Are the benchmark results stable or do they have a large variance?

I ran `DaCapo` twice, but with `Renaissance-jmh` it was just a single run. 

>Did you also compare against the version without your changes? That should be done for sanity checking. You have modified the code for the legacy mode, too.
>I'm not sure what you mean by "With/Without fastLock"? Patch applied? LockingMode selection?

Patch was applied, but for in first case LockingMode was not selected, and in second case (with fast locking) `LockingMode=2` argument was given. 

But I will run these again with/without the code change. Thanks for insights.