@TheRealMDoerr 2023-06-26
> After struggling a lot with my environment, I finally was able to run some performance tests. There is the "db" subtest of JVM98 which caught my attention. It reproducibly shows
> 
> * LockingMode=0: avg time: 2888ms
> * LockingMode=1: avg time: 867ms
> * LockingMode=2: avg time: 1020ms
> 
> We need to gain deeper insight into the why before integration.

Note that the new fast locking (LockingMode 2) doesn't support fast recursive locking. However, I doubt that this is the problem, here. LockingMode 2 is fastest on Power10 for "db" benchmark. So, yes, interesting finding!