@offamitkumar 2023-06-26
with 

fastlockbench:

|                 | Time (ns/op) [Average] | Performance Improvement |
|-----------------|-----------------------|-------------------------|
| With Patch      | 18.614                | -                       |
| Without Patch (+LockingMode=2)   | 19.116                | -                       |
| Improvement     | -                     | 2.63%                   |


Dacapo:

|                 | Time (msec)            | Performance Improvement |
|-----------------|-----------------------|-------------------------|
| Without Patch   | 118199                | -                       |
| With Patch (+LockingMode=2)      | 120213                | -                       |
| Improvement     | -                     | -1.70%                  |



 Renaissance-jmh:
| Benchmark                                           | Without Patch | With Patch (+LockingMode=2) | Improvement (%) |
|-----------------------------------------------------|--------------|-------------|-----------------|
| o.r.actors.JmhAkkaUct.runOperation                  | 1349.563     | 1456.927    | -8.00           |
| o.r.actors.JmhReactors.runOperation                 | 10296.818    | 11605.138   | -12.82          |
| o.r.jdk.concurrent.JmhFjKmeans.runOperation         | 1145.668     | 1390.726    | -21.35          |
| o.r.jdk.concurrent.JmhFutureGenetic.runOperation    | 1952.648     | 1976.534    | -1.22           |
| o.r.jdk.streams.JmhParMnemonics.runOperation        | 2819.193     | 2801.261    | 0.64            |
| o.r.jdk.streams.JmhScrabble.runOperation            | 78.594       | 74.529      | 5.16            |
| o.r.rx.JmhRxScrabble.runOperation                   | 171.220      | 180.978     | -5.70           |
| o.r.scala.sat.JmhScalaDoku.runOperation             | 5841.122     | 3556.529    | 39.15           |
| o.r.scala.stdlib.JmhScalaKmeans.runOperation        | 199.683      | 183.357     | 8.18            |
| o.r.scala.stm.JmhPhilosophers.runOperation          | 14482.716    | 15834.972   | -9.32           |
| o.r.scala.stm.JmhScalaStmBench7.runOperation        | 1567.814     | 1716.439    | -9.46           |
| o.r.twitter.finagle.JmhFinagleChirper.runOperation  | 9477.834     | 9737.119    | -2.73           |
| o.r.twitter.finagle.JmhFinagleHttp.runOperation     | 4381.338     | 4681.571    | -6.85           |


These are the new results I'm getting, seems consistent as well. `-21.35 ` appears to be a lot but not sure if it's of observable category or not.