@coleenp 2023-09-07
We are considering making Fast Locking on by default for Oracle supported platforms.  Have these performance concerns been addressed?