@RealLucy 2023-09-08
> We are considering making Fast Locking on by default for Oracle supported platforms. Have these performance concerns been addressed?

The performance regression which is observed in some tests is still not fully understood. I will approve the PR despite of that. According to all out testing, it is functionally correct.